requirements.txt
django-environ==0.4.5
django==3.2.3
djangorestframework==3.12.4
psycopg2-binary==2.8.6
amqp==5.0.6
celery==5.1.0
redis==3.5.3
flower==0.9.5
requests==2.25.1
gunicorn==20.1.0
django_extensions==3.1.3